# openapi.yaml
openapi: 3.1.0
info:
  title: Execution Plane API
  description: Accept, track, and retrieve execution work independently of Syntara
    workflows.
  version: 1.0.0
paths:
  /v1/capabilities:
    get:
      summary: Capabilities
      operationId: capabilities_v1_capabilities_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CapabilitiesResponse'
  /v1/work-items:
    post:
      summary: Submit Work Item
      operationId: submit_work_item_v1_work_items_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkItemSubmit'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Work Items
      operationId: list_work_items_v1_work_items_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkItemRead'
                title: Response List Work Items V1 Work Items Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/execution-targets:
    get:
      summary: List Execution Targets
      operationId: list_execution_targets_v1_execution_targets_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExecutionTargetRead'
                title: Response List Execution Targets V1 Execution Targets Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/cluster-bindings/{source_integration_id}:
    put:
      summary: Upsert Cluster Binding
      description: Persist a versioned desired state for the independently managed
        cluster.
      operationId: upsert_cluster_binding_v1_cluster_bindings__source_integration_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: source_integration_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Integration Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterBindingUpsert'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterBindingRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Get Cluster Binding
      description: Return safe desired and observed state for a source integration.
      operationId: get_cluster_binding_v1_cluster_bindings__source_integration_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: source_integration_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Source Integration Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterBindingRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/work-items/{work_id}:
    get:
      summary: Get Work Item
      operationId: get_work_item_v1_work_items__work_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: work_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Work Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/work-items/{work_id}/cancel:
    post:
      summary: Cancel Work Item
      operationId: cancel_work_item_v1_work_items__work_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: work_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Work Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkItemRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BackendType:
      type: string
      enum:
      - vanilla_k8s
      - openshell
      title: BackendType
      description: Supported execution target backends.
    CapabilitiesResponse:
      properties:
        api_version:
          type: string
          title: Api Version
          default: v1
        workloads:
          items:
            type: string
          type: array
          title: Workloads
        result_events:
          type: boolean
          title: Result Events
          default: true
      type: object
      title: CapabilitiesResponse
      description: Contract and workload features implemented by this service release.
    ClusterBindingRead:
      properties:
        client_id:
          type: string
          title: Client Id
        source_integration_id:
          type: string
          format: uuid
          title: Source Integration Id
        desired_revision:
          type: integer
          title: Desired Revision
        observed_revision:
          type: integer
          title: Observed Revision
        cluster_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Cluster Id
        status:
          type: string
          title: Status
        status_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Message
        enabled:
          type: boolean
          title: Enabled
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - client_id
      - source_integration_id
      - desired_revision
      - observed_revision
      - cluster_id
      - status
      - status_message
      - enabled
      - updated_at
      title: ClusterBindingRead
      description: Safe observed state for an AO-owned source integration.
    ClusterBindingUpsert:
      properties:
        revision:
          type: integer
          minimum: 1.0
          title: Revision
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        endpoint:
          type: string
          maxLength: 2048
          minLength: 1
          title: Endpoint
        namespace:
          type: string
          maxLength: 63
          minLength: 1
          pattern: ^[a-z0-9]([-a-z0-9]*[a-z0-9])?$
          title: Namespace
        credential:
          type: string
          title: Credential
          default: ''
        ca_certificate:
          anyOf:
          - type: string
            maxLength: 65536
          - type: 'null'
          title: Ca Certificate
        insecure_skip_tls_verify:
          type: boolean
          title: Insecure Skip Tls Verify
          default: false
        project_ids:
          anyOf:
          - items:
              type: string
              format: uuid
            type: array
          - type: 'null'
          title: Project Ids
        labels:
          additionalProperties:
            type: string
          type: object
          title: Labels
        enabled:
          type: boolean
          title: Enabled
          default: true
      type: object
      required:
      - revision
      - name
      - endpoint
      - namespace
      title: ClusterBindingUpsert
      description: Versioned OpenShift integration desired state supplied by an authorized
        client.
    ExecutionTargetRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        name:
          type: string
          title: Name
        backend_type:
          $ref: '#/components/schemas/BackendType'
        endpoint:
          type: string
          title: Endpoint
        placement:
          oneOf:
          - $ref: '#/components/schemas/KubernetesPlacement'
          - $ref: '#/components/schemas/RHELPlacement'
          title: Placement
          discriminator:
            propertyName: type
            mapping:
              kubernetes: '#/components/schemas/KubernetesPlacement'
              rhel: '#/components/schemas/RHELPlacement'
        status:
          $ref: '#/components/schemas/TargetStatus'
        enabled:
          type: boolean
          title: Enabled
        is_default:
          type: boolean
          title: Is Default
        status_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Message
        labels:
          additionalProperties:
            type: string
          type: object
          title: Labels
        created_at:
          type: string
          format: date-time
          title: Created At
        last_ran_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Ran At
      type: object
      required:
      - id
      - cluster_id
      - name
      - backend_type
      - endpoint
      - placement
      - status
      - enabled
      - is_default
      - status_message
      - labels
      - created_at
      - last_ran_at
      title: ExecutionTargetRead
      description: Safe execution-target representation without management credentials.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    KubernetesPlacement:
      properties:
        type:
          type: string
          const: kubernetes
          title: Type
          default: kubernetes
        namespace:
          type: string
          maxLength: 63
          minLength: 1
          pattern: ^[a-z0-9]([-a-z0-9]*[a-z0-9])?$
          title: Namespace
          description: Kubernetes namespace reserved for execution workloads
        node_selectors:
          items:
            type: string
          type: array
          title: Node Selectors
        tolerations:
          items:
            type: string
          type: array
          title: Tolerations
      type: object
      required:
      - namespace
      title: KubernetesPlacement
      description: Metadata required to place workers on Kubernetes.
    RHELPlacement:
      properties:
        type:
          type: string
          const: rhel
          title: Type
          default: rhel
      type: object
      title: RHELPlacement
      description: Placeholder for RHEL-specific execution-target metadata.
    TargetStatus:
      type: string
      enum:
      - registering
      - validating
      - bootstrapping
      - active
      - degraded
      - draining
      - failed
      title: TargetStatus
      description: Lifecycle states of an execution target.
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkItemRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          $ref: '#/components/schemas/WorkItemStatus'
        result:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Result
        created_at:
          type: string
          format: date-time
          title: Created At
        claimed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Claimed At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        resource_cleanup_status:
          type: string
          title: Resource Cleanup Status
        resource_cleanup_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Cleanup Error
        completion_event_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Completion Event Id
        state_revision:
          anyOf:
          - type: integer
          - type: 'null'
          title: State Revision
      type: object
      required:
      - id
      - status
      - result
      - created_at
      - claimed_at
      - completed_at
      - resource_cleanup_status
      - resource_cleanup_error
      title: WorkItemRead
      description: Safe work-item representation; excludes credentials and persistence
        internals.
    WorkItemStatus:
      type: string
      enum:
      - pending
      - claimed
      - dispatched
      - cancel_requested
      - reconciliation_required
      - completed
      - failed
      - cancelled
      title: WorkItemStatus
      description: Lifecycle states of a dispatched work item.
    WorkItemSubmit:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workload_type:
          type: string
          const: script
          title: Workload Type
          default: script
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - id
      - payload
      title: WorkItemSubmit
      description: Script work accepted from a trusted client.
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer